Add lighten and darken functions to colors
