+++ app/helpers/colors_helper.rb
@@ -49,15 +49,6 @@ module ColorsHelper
     colored_thing.color_id
   end
 
-  def darken_color(hex_color, amount = 0.4)
-    hex_color = hex_color.delete('#')
-    rgb = hex_color.scan(/../).map(&:hex)
-    rgb[0] = (rgb[0].to_i * amount).round
-    rgb[1] = (rgb[1].to_i * amount).round
-    rgb[2] = (rgb[2].to_i * amount).round
-    "#%02x%02x%02x" % rgb
-  end
-
   def colored_text(color)
     background = color.contrasting_color(dark_color: '#333', light_color: 'transparent')
     style = "background-color: #{background}; color: #{color.hexcode}"
